Format_Agent/format_agent/pyproject.toml
[project]
name = "format-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.8.0",
    "openai-agents>=0.3.0",
]
